[EC-150] Refactoring: Moved mocks to new Mocks object
