+++ src/test/scala/io/iohk/ethereum/ledger/BlockRewardSpec.scala
@@ -1,6 +1,6 @@
 package io.iohk.ethereum.ledger
 
-import io.iohk.ethereum.Fixtures
+import io.iohk.ethereum.{Fixtures, Mocks}
 import io.iohk.ethereum.db.components.{SharedEphemDataSources, Storages}
 import io.iohk.ethereum.domain.{Account, Address, Block}
 import io.iohk.ethereum.utils.Config
@@ -8,7 +8,7 @@ import org.scalatest.{FlatSpec, Matchers}
 
 class BlockRewardSpec extends FlatSpec with Matchers {
 
-  val ledger = new LedgerImpl(new Fixtures.MockVM())
+  val ledger = new LedgerImpl(new Mocks.MockVM())
 
   "Reward Calculation" should "pay to the miner if no ommers included" in new TestSetup {
     val block = sampleBlock(validAccountAddress, Seq(validAccountAddress2, validAccountAddress3))
